Add sliding-window sync finder
Sometimes there are gaps in the file data which can't be accounted for
by padding offsets. Add a sliding-window algorithm to find the next
valid file-header.

This can also be useful if a tape has been spliced to repair a break. If
the break is mid-file, you may need to edit the tape data to
deliberately corrupt the file's FID header, or the next file may be
skipped.